requirements.txt
streamlit
llama-index
llama-index-llms-google-genai
llama-index-embeddings-google-genai
llama-index-vector-stores-chroma
chromadb
python-dotenv
pypdf

